requirements.txt
webuiapi
mypy
pylint
Pillow
types-Pillow
types-requests